AK: Move memory streams into their own header.
